pyproject.toml
[project]
name = "rhizome"
version = "0.1.0"
description = "Rhizome（灵犀树）- 基于LangChain的知识体系智能体"
readme = "SPEC.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.1.0",
    "openai>=1.10.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "huggingface-hub>=0.20.0,<1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sentence-transformers>=5.3.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knowledge_agent", "backend"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
